Updated attribute handling (from Percy)
The hints for Attributes related to an effect ("Cynet Codec", "Chaofeng, Phantom of the Yang Zing") will now use the system string and a custom utility that matches an attribute(s) will all corresponding strings. A function for Races has been added for completion's sake, but it's not currently used.