Change for "Cursed Bamboo Sword"
Changed the operation such that the script won't error anymore if you bounce the monster the card is equipped to. Although there are no official Bamboo Sword monsters, the script doesn't change a single thing when bouncing a different card. The changes are also tested thouroughly before opening this PR and I asked Hel whether this is even feasable. No objections from them, so I think I can give it a try.