"Plunder Patroll Booty" fix
Ghost Belle should be able to negate the activation of its first effect and changing the Attribute of the target and Special Summoning/returning to the Deck the "Plunder Patroll" monster do not happen at the same time. Also now shows the selected monster to the opponent if the player chooses to return it to the Deck.