Update counter traps that respond to summons (#664)
Update their way of checking the current chain to account for cards copying their effects, before the cards copying them would call their conditions in the target, and this would return false as Duel.GetCurrentChain would have returned a chain count of 1. Use the new parameter to make it return the proper count when being checked in the cost/target.