"Puppet King" + "Puppet Queen" fix
Added a new custom event so that their effects cannot be activated when a monster is added to the hand by an effect that doesn't reveal the card (e.g. "Bingo Machine, Go!!!"). In other words, the game state from the player's perspective doesn't know that the card added to the hand is a monster. Also, general script polish for "Puppet King" & "Puppet Queen".