Fix Scapegoat (GOAT)
Previously, the script only checked for the count of summons the player did at the end of a chain, in regards to being activated as Chain Link 2 or higher. Now a check has been added to count summons outside of chains, for when the card would be activated as Chain Link 2 or higher.