Fix to Reptilianne Melusine
Changed the way the card tracks the materials used for its summon. It now uses a flag system, so when it is flipped face-down it will properly "forget" the type of monsters used for its summon and will not retain the indestructibility effect. This should also clear up the issue with Rebirth Judgment incorrectly removing the indestructibility effect if it changes the monster type of cards in the graveyard to anything other than reptile.

The flag is using string id 3, so something should be added to the cdb to give the player more info, like: "Summoned using only Reptiles."