"Prologue of the Destruction Swordsman" fix
Was always checking if the player had an empty zone from the Extra Deck for the monster to be summoned, making the player unable to choose a "Buster Dragon" in the GY in some scenarios + shouldn't be able to use its GY effect if it has already successfully resolved earlier in the same turn. Also includes general script polishing and updates.