Update "Rule of the Day"
The "Rule of the Day" duel mode will apply the following rules:

At the start of each Draw Phase: Each player declares Spell or Trap Cards. Add from the Worlds banlist 1 random card of that type to that player's hand. If the card is a Trap card, it can be Set instead, and it can be activated this turn.