Simplify redundant logic regarding the ternary 'or'
There were many cases of the pattern `x and x or y`.
Due to operator precedence this is evaluated as `(x and x) or y`.
Since `x and x` is equivalent to `x` itself this can be simplified,
improving clarity and sometimes removing unnecessary function calls.
(don't know if Lua would optimize them anyways though)