"Duel.Overlay" update
Overwrite the function to add handling for when the Xyz Monster itself or any of the cards to be attached to it are unaffected by the currently resolving effect according to rulings. Non-activated effects that attach a card to an Xyz Monster will still need to check for effect immunity manually.