Small Fixes "Icejade Aegirine Gymir" and "S:P Little Knight"
Icejade Aegirine Gymir:
Now the effect that prevents banishing by card effects applies whether the reason for banishing was temporary or permanent.
S:P Little Knight:
Now the effect that banishes 2 monsters temporarily checks when the effect resolves if both can be banished, for the cases that only 1 still can be.