Updated cards with multi-choice effects that also have a cost
After a recent ruling change you first pay the cost and then choose which effect to activate. If paying the cost makes an effect activatable then you can choose to activate that effect, however you still need to be able to activate at least 1 of the listed effects in order to begin the activation procedure in the first place. Some example cards affected by this change are: "Daigusto Emeral", "Number 60: Dugares the Timeless", and "Meliae of the Trees".