"Normal/Flip/Special Summoned this turn" update
A monster that stopped being face-up on the field (e.g. due to being flipped face-down or banished temporarily) shouldn't be treated as a monster that was Normal, Flip, or Specal Summoned that turn. Also, in the case of being banished temporarily it shouldn't be treated as a monster that has already changed its battle position this turn. Updated some of the generic functions that those cards tend to use to handle those scenario.